modernize and lock SkbKeyring
- As I'm going through in an attempt to repair the LKsec breakage, I saw that the SKBKeyring class could stand a fair amount of improvement. This PR:
  - contextifies SkbKeyring so we're not using global context G
  - locks access to the class so that we won't get races across go routines
