+++ go/libkb/keyring.go
@@ -57,7 +57,7 @@ func LoadSKBKeyring(un NormalizedUsername, g *GlobalContext) (*SKBKeyringFile, e
 		return nil, NoUsernameError{}
 	}
 
-	skbfile := NewSKBKeyringFile(g.SKBFilenameForUser(un))
+	skbfile := NewSKBKeyringFile(g, g.SKBFilenameForUser(un))
 	err := skbfile.LoadAndIndex()
 	if err != nil && !os.IsNotExist(err) {
 		return nil, err
@@ -137,8 +137,8 @@ func (k KeyringFile) WriteTo(w io.Writer) (int64, error) {
 
 func (k KeyringFile) GetFilename() string { return k.filename }
 
-func (k KeyringFile) Save() error {
-	return SafeWriteToFile(k, 0)
+func (k KeyringFile) Save(g *GlobalContext) error {
+	return SafeWriteToFile(g.Log, k, 0)
 }
 
 type SecretKeyType int
